requireAdmin function

+++ api/productRouter.js
@@ -10,6 +10,7 @@ const {
 const client = require("../db/index");
 const productRouter = express.Router();
 productRouter.use(express.json());
+const { requireUser, requireAdmin } = require("./utils");
 
 //get all
 // Tested with postman and is working
@@ -32,30 +33,34 @@ productRouter.get("/:productId", async (req, res) => {
 
 //create a product - admin only
 // Tested with postman and is working
-productRouter.post("/", async (req, res, next) => {
+productRouter.post("/", requireAdmin, async (req, res, next) => {
   const { title, designer, description, price, category, image, count } =
     req.body;
 
-  try {
-    const response = await createProduct({
-      title,
-      designer,
-      description,
-      price,
-      category,
-      image,
-      count,
-    });
+  if (!req.admin) {
+    res.send({ error: "Must be admin" });
+  } else {
+    try {
+      const response = await createProduct({
+        title,
+        designer,
+        description,
+        price,
+        category,
+        image,
+        count,
+      });
 
-    res.send(response);
-  } catch (err) {
-    res.send({ error: err.message });
+      res.send(response);
+    } catch (err) {
+      res.send({ error: err.message });
+    }
   }
 });
 
 // update a product - admin only
 // Tested with postman and is working
-productRouter.patch("/:productId", async (req, res, next) => {
+productRouter.patch("/:productId", requireAdmin, async (req, res, next) => {
   const { productId: id } = req.params;
 
   const { title, designer, description, price, category, count } = req.body;
@@ -70,24 +75,34 @@ productRouter.patch("/:productId", async (req, res, next) => {
     count,
   };
 
-  try {
-    const updatedProduct = await updateProduct(updateFields);
-    res.send(updatedProduct);
-  } catch (err) {
-    throw error;
+  if (!req.admin) {
+    res.send({ error: "Must be admin" });
+  } else {
+    try {
+      const updatedProduct = await updateProduct(updateFields);
+      res.send(updatedProduct);
+    } catch (err) {
+      throw error;
+    }
   }
 });
 
 // delete a product - admin only
 // Tested with postman and is working
-productRouter.delete("/:productId", async (req, res, next) => {
+productRouter.delete("/:productId", requireAdmin, async (req, res, next) => {
   const { productId: id } = req.params;
-  try {
-    const deletedProduct = await destroyProduct(id);
-    res.send({ message: "Product deleted" });
-    return;
-  } catch (err) {
-    next(err);
+
+  if (!req.admin) {
+    res.send({ error: "Must be admin" });
+  } else {
+    try {
+      console.log(req.admin);
+      const deletedProduct = await destroyProduct(id);
+      res.send({ message: "Product deleted" });
+      return;
+    } catch (err) {
+      next(err);
+    }
   }
 });
